+++ src/jtree.ts
@@ -23,7 +23,7 @@ class jtree {
   static GrammarProgram = GrammarProgram
   static UnknownGrammarProgram = UnknownGrammarProgram
   static TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode
-  static getVersion = () => "26.2.0"
+  static getVersion = () => "26.3.0"
 }
 
 export default jtree